Revert "atomic: cast to unsigned when shifting"
This reverts commit c346481f4bcc5aa52662e4dffa2c17d7b5b18509.

+++ common/button.c
@@ -62,7 +62,7 @@ static atomic_t sim_button_state;
  */
 static int simulated_button_pressed(const struct button_config *button)
 {
-	return !!((uint32_t)sim_button_state & BIT(button->type));
+	return !!(sim_button_state & BIT(button->type));
 }
 #endif
 
@@ -389,8 +389,7 @@ static void simulate_button_release_deferred(void)
 	/* Release the button */
 	for (button_idx = 0; button_idx < BUTTON_COUNT; button_idx++) {
 		/* Check state for button pressed */
-		if ((uint32_t)sim_button_state &
-		    BIT(buttons[button_idx].type)) {
+		if (sim_button_state & BIT(buttons[button_idx].type)) {
 			/* Set state of the button as released */
 			atomic_clear_bits(&sim_button_state,
 					  BIT(buttons[button_idx].type));
